Refactor: 코드리뷰 후 수정사항 반영
