+++ backend/src/main/java/com/api/backend/category/data/entity/ScheduleCategory.java
@@ -3,7 +3,7 @@ package com.api.backend.category.data.entity;
 import com.api.backend.category.data.dto.ScheduleCategoryEditRequest;
 import com.api.backend.category.type.CategoryType;
 import com.api.backend.global.domain.BaseEntity;
-import com.api.backend.schedule.data.enetity.Schedule;
+import com.api.backend.schedule.data.entity.Schedule;
 import com.api.backend.team.data.entity.Team;
 import java.util.ArrayList;
 import java.util.List;
