+++ backend/src/main/java/com/api/backend/schedule/data/dto/ScheduleResponse.java
@@ -1,8 +1,6 @@
 package com.api.backend.schedule.data.dto;
 
-import com.api.backend.schedule.data.enetity.TeamParticipantsSchedule;
 import com.api.backend.schedule.data.type.RepeatCycle;
-import com.api.backend.team.data.entity.TeamParticipants;
 import com.api.backend.team.data.type.TeamRole;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
@@ -11,6 +9,7 @@ import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import org.springframework.data.domain.Page;
 
 @Getter
 @Builder
@@ -31,6 +30,7 @@ public class ScheduleResponse {
   private List<Long> teamParticipantsIds;
   private List<String> teamParticipantsNames;
   private List<TeamRole> teamRoles;
+
   public static ScheduleResponse from(ScheduleDto scheduleDto) {
     return ScheduleResponse.builder()
         .scheduleId(scheduleDto.getScheduleId())
@@ -48,14 +48,7 @@ public class ScheduleResponse {
         .build();
   }
 
-  public static List<ScheduleResponse> from(List<ScheduleDto> scheduleDtoList) {
-    if (scheduleDtoList != null) {
-      List<ScheduleResponse> scheduleResponses = new ArrayList<>();
-      for (ScheduleDto dto : scheduleDtoList) {
-        scheduleResponses.add(from(dto));
-      }
-      return scheduleResponses;
-    }
-    return new ArrayList<>();
+  public static Page<ScheduleResponse> from(Page<ScheduleDto> scheduleDtoList) {
+    return scheduleDtoList.map(ScheduleResponse::from);
   }
 }
