+++ backend/src/main/java/com/api/backend/schedule/data/repository/ScheduleRepository.java
@@ -1,10 +1,12 @@
 package com.api.backend.schedule.data.repository;
 
-import com.api.backend.schedule.data.enetity.Schedule;
+import com.api.backend.schedule.data.entity.Schedule;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ScheduleRepository extends JpaRepository<Schedule,Long> {
-
+  Page<Schedule> findAllByTeam_TeamId(Pageable pageable, Long teamId);
 }
